refactor(filepicker): split oversized file browser/import screens
- EnhancedFileBrowser.kt:        1346 -> 733 (10 sibling files)
- EnhancedStorageBrowserScreen.kt: 962 -> 572 (5 sibling files)
- FolderImportScreen.kt:           959 -> 652 (4 sibling files)

File items, dialogs, navigation chrome, and import option cards each
move to their own sibling files in the same package.
