+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/filepicker/EnhancedStorageBrowserScreen.kt
@@ -418,242 +418,7 @@ fun EnhancedStorageBrowserScreen(
     }
 }
 
-@Composable
-private fun EnhancedFileItem(
-    file: File,
-    viewMode: ViewMode,
-    isSelected: Boolean,
-    isSelectable: Boolean,
-    onClick: () -> Unit,
-    onLongClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        modifier = modifier
-            .fillMaxWidth()
-            .then(
-                if (isSelectable) {
-                    Modifier.selectable(
-                        selected = isSelected,
-                        role = Role.Checkbox,
-                        onClick = onClick
-                    )
-                } else {
-                    Modifier.clickable(onClick = onClick)
-                }
-            )
-            .then(
-                if (isSelected) {
-                    Modifier.background(
-                        MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
-                    )
-                } else {
-                    Modifier
-                }
-            ),
-        shape = RoundedCornerShape(12.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            horizontalArrangement = Arrangement.spacedBy(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Checkbox for multi-selection
-            if (isSelectable) {
-                Checkbox(
-                    checked = isSelected,
-                    onCheckedChange = { onClick() },
-                    modifier = Modifier.size(24.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-            }
-            
-            // File/Folder icon
-            FileIcon(file = file)
 
-            // File info
-            Column(
-                modifier = Modifier.weight(1f),
-                verticalArrangement = Arrangement.spacedBy(4.dp)
-            ) {
-                Text(
-                    text = file.name,
-                    style = MaterialTheme.typography.bodyLarge,
-                    fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Medium,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-                )
-
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    if (file.isFile) {
-                        Text(
-                            text = formatFileSize(file.length()),
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    } else {
-                        val itemCount = try {
-                            file.listFiles()?.size ?: 0
-                        } catch (e: SecurityException) {
-                            0
-                        }
-                        Text(
-                            text = "$itemCount items",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-
-                    Text(
-                        text = formatDate(file.lastModified()),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-
-            // Chevron for folders or checkmark for selected
-            when {
-                file.isDirectory -> {
-                    Icon(
-                        Icons.Default.ChevronRight,
-                        contentDescription = "Media image",
-                        tint = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-                isSelected -> {
-                    Icon(
-                        Icons.Default.CheckCircle,
-                        contentDescription = "Selected",
-                        tint = MaterialTheme.colorScheme.primary
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ImportConfirmationDialog(
-    selectedFiles: List<File>,
-    selectedFolders: List<File>,
-    onConfirm: (includeSubfolders: Boolean, validateFiles: Boolean) -> Unit,
-    onDismiss: () -> Unit
-) {
-    var includeSubfolders by remember { mutableStateOf(true) }
-    var validateFiles by remember { mutableStateOf(true) }
-    
-    val totalFiles = selectedFiles.size
-    val totalFolders = selectedFolders.size
-    val estimatedFiles = totalFiles + (totalFolders * 50) // Rough estimate
-    
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { 
-            Text("Import Selection") 
-        },
-        text = {
-            Column(
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                Text(
-                    text = "You've selected:",
-                    style = MaterialTheme.typography.bodyMedium
-                )
-                
-                // Selection summary
-                Card(
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.surfaceVariant
-                    )
-                ) {
-                    Column(
-                        modifier = Modifier.padding(12.dp),
-                        verticalArrangement = Arrangement.spacedBy(4.dp)
-                    ) {
-                        Text("• $totalFiles files")
-                        Text("• $totalFolders folders")
-                        if (includeSubfolders) {
-                            Text(
-                                "• Estimated ~$estimatedFiles total files (including subfolders)",
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-                    }
-                }
-                
-                // Options
-                if (totalFolders > 0) {
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Column(modifier = Modifier.weight(1f)) {
-                            Text(
-                                text = "Include subfolders",
-                                style = MaterialTheme.typography.bodyMedium
-                            )
-                            Text(
-                                text = "Recursively add all files in subfolders",
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-                        Switch(
-                            checked = includeSubfolders,
-                            onCheckedChange = { includeSubfolders = it }
-                        )
-                    }
-                }
-                
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            text = "Validate file types",
-                            style = MaterialTheme.typography.bodyMedium
-                        )
-                        Text(
-                            text = "Only import supported media files",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                    Switch(
-                        checked = validateFiles,
-                        onCheckedChange = { validateFiles = it }
-                    )
-                }
-            }
-        },
-        confirmButton = {
-            Button(
-                onClick = {
-                    onConfirm(includeSubfolders, validateFiles)
-                }
-            ) {
-                Icon(Icons.Default.ImportExport, contentDescription = "Import or export")
-                Spacer(modifier = Modifier.width(8.dp))
-                Text("Import")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        }
-    )
-}
 
 private fun getAllFilesRecursively(
     folder: File, 
@@ -689,58 +454,6 @@ private fun getFilesInFolder(
 }
 
 // Reuse existing components from original StorageBrowserScreen
-@Composable
-private fun QuickAccessBar(
-    onPathSelected: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val externalStorage = Environment.getExternalStorageDirectory().absolutePath
-    val downloadsDir = Environment
-        .getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
-        .absolutePath
-    val documentsDir = Environment
-        .getExternalStoragePublicDirectory(Environment.DIRECTORY_DOCUMENTS)
-        .absolutePath
-    val dcimDir = Environment
-        .getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM)
-        .absolutePath
-    val musicDir = Environment
-        .getExternalStoragePublicDirectory(Environment.DIRECTORY_MUSIC)
-        .absolutePath
-
-    Row(
-        modifier = modifier
-            .padding(8.dp)
-            .horizontalScroll(rememberScrollState()),
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        QuickAccessChip(
-            label = "Internal",
-            icon = Icons.Default.PhoneAndroid,
-            onClick = { onPathSelected(externalStorage) }
-        )
-        QuickAccessChip(
-            label = "Downloads",
-            icon = Icons.Default.Download,
-            onClick = { onPathSelected(downloadsDir) }
-        )
-        QuickAccessChip(
-            label = "Documents",
-            icon = Icons.Default.Folder,
-            onClick = { onPathSelected(documentsDir) }
-        )
-        QuickAccessChip(
-            label = "DCIM",
-            icon = Icons.Default.PhotoLibrary,
-            onClick = { onPathSelected(dcimDir) }
-        )
-        QuickAccessChip(
-            label = "Music",
-            icon = Icons.Default.MusicNote,
-            onClick = { onPathSelected(musicDir) }
-        )
-    }
-}
 
 @Composable
 private fun QuickAccessChip(
@@ -761,38 +474,6 @@ private fun QuickAccessChip(
     )
 }
 
-@Composable
-private fun FileIcon(file: File) {
-    val (icon, gradient) = when {
-        file.isDirectory -> Icons.Default.Folder to listOf(Color(0xFFFFA726), Color(0xFFFF6F00))
-        file.extension.lowercase() in listOf("epub", "pdf", "txt", "mobi") -> 
-            Icons.Default.Book to listOf(Color(0xFF1B5E20), Color(0xFF4CAF50))
-        file.extension.lowercase() in listOf("mp3", "m4a", "flac", "wav") -> 
-            Icons.Default.MusicNote to listOf(Color(0xFF4A148C), Color(0xFF9C27B0))
-        file.extension.lowercase() in listOf("mp4", "mkv", "avi", "mov") -> 
-            Icons.Default.Movie to listOf(Color(0xFF0D47A1), Color(0xFF2196F3))
-        file.extension.lowercase() in listOf("cbz", "cbr", "cb7") -> 
-            Icons.Default.AutoStories to listOf(Color(0xFFE65100), Color(0xFFFF9800))
-        file.extension.lowercase() in listOf("jpg", "jpeg", "png", "gif") -> 
-            Icons.Default.Image to listOf(Color(0xFF006064), Color(0xFF00BCD4))
-        else -> Icons.AutoMirrored.Filled.InsertDriveFile to listOf(Color(0xFF37474F), Color(0xFF78909C))
-    }
-
-    Box(
-        modifier = Modifier
-            .size(48.dp)
-            .clip(CircleShape)
-            .background(Brush.linearGradient(gradient)),
-        contentAlignment = Alignment.Center
-    ) {
-        Icon(
-            icon,
-            contentDescription = "Media image",
-            tint = Color.White,
-            modifier = Modifier.size(24.dp)
-        )
-    }
-}
 
 @Composable
 private fun EmptyFolderView() {
@@ -872,77 +553,6 @@ private fun ErrorView(
     }
 }
 
-@Composable
-private fun StoragePermissionCard(
-    onRequestPermission: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Box(
-        modifier = modifier,
-        contentAlignment = Alignment.Center
-    ) {
-        Card(
-            modifier = Modifier
-                .padding(24.dp)
-                .fillMaxWidth(),
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surfaceVariant
-            )
-        ) {
-            Column(
-                modifier = Modifier.padding(24.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Icon(
-                    Icons.Default.Folder,
-                    contentDescription = "Media image",
-                    modifier = Modifier.size(72.dp),
-                    tint = MaterialTheme.colorScheme.primary
-                )
-                
-                Text(
-                    text = "Storage Access Required",
-                    style = MaterialTheme.typography.titleLarge,
-                    fontWeight = FontWeight.Bold,
-                    textAlign = TextAlign.Center
-                )
-                
-                Text(
-                    text = "To browse and open ebooks, documents, and other files from your device storage, CleverFerret needs 'All Files Access' permission.",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    textAlign = TextAlign.Center
-                )
-                
-                Text(
-                    text = "This permission allows the app to read files like EPUB, PDF, MOBI, and other document formats stored on your device.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    textAlign = TextAlign.Center
-                )
-                
-                Spacer(modifier = Modifier.height(8.dp))
-                
-                Button(
-                    onClick = onRequestPermission,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Icon(Icons.Default.Security, contentDescription = "Security")
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Text("Grant Access")
-                }
-                
-                Text(
-                    text = "You'll be taken to system settings. Find 'CleverFerret' and enable 'Allow access to manage all files'.",
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.7f),
-                    textAlign = TextAlign.Center
-                )
-            }
-        }
-    }
-}
 
 private fun formatFileSize(bytes: Long): String {
     if (bytes <= 0) return "0 B"
