+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/folderimport/FolderImportScreen.kt
@@ -493,170 +493,8 @@ fun FolderImportScreen(
     }
 }
 
-@Composable
-private fun UrlImportDialog(
-    url: String,
-    onUrlChange: (String) -> Unit,
-    onDismiss: () -> Unit,
-    onImport: () -> Unit
-) {
-    val isValidUrl = remember(url) {
-        runCatching {
-            val parsed = URI(url.trim())
-            val scheme = parsed.scheme?.lowercase()
-            (scheme == "http" || scheme == "https") && !parsed.path.isNullOrBlank()
-        }.getOrDefault(false)
-    }
 
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Import from URL") },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                OutlinedTextField(
-                    value = url,
-                    onValueChange = onUrlChange,
-                    label = { Text("Media URL") },
-                    singleLine = true,
-                    isError = url.isNotBlank() && !isValidUrl
-                )
-                Text(
-                    text = "Direct links only (http/https) ending in supported extensions, e.g. .mp3, .epub, .pdf, .mp4.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        },
-        confirmButton = {
-            TextButton(
-                onClick = onImport,
-                enabled = isValidUrl
-            ) { Text("Import") }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) { Text("Cancel") }
-        }
-    )
-}
 
-@Composable
-private fun ImportOptionCard(
-    icon: ImageVector,
-    title: String,
-    description: String,
-    onClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceVariant
-        )
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(20.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Surface(
-                shape = CircleShape,
-                color = MaterialTheme.colorScheme.primary.copy(alpha = 0.15f),
-                modifier = Modifier.size(56.dp)
-            ) {
-                Icon(
-                    icon,
-                    contentDescription = null,
-                    modifier = Modifier
-                        .padding(14.dp)
-                        .fillMaxSize(),
-                    tint = MaterialTheme.colorScheme.primary
-                )
-            }
-            
-            Spacer(modifier = Modifier.width(16.dp))
-            
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.SemiBold
-                )
-                Text(
-                    text = description,
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-            
-            Icon(
-                Icons.Default.ChevronRight,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-}
-
-@Composable
-private fun SupportedFormatsCard() {
-    Card(
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surface
-        ),
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Text(
-                text = "Supported Formats",
-                style = MaterialTheme.typography.titleSmall,
-                fontWeight = FontWeight.SemiBold
-            )
-            
-            Spacer(modifier = Modifier.height(12.dp))
-            
-            FormatRow(
-                icon = Icons.AutoMirrored.Filled.MenuBook,
-                label = "Books",
-                formats = "EPUB, PDF, MOBI, AZW, FB2, TXT, RTF, DJVU"
-            )
-            
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            FormatRow(
-                icon = Icons.Default.Image,
-                label = "Comics",
-                formats = "CBZ, CBR, CB7, PDF"
-            )
-            
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            FormatRow(
-                icon = Icons.Default.MusicNote,
-                label = "Audio",
-                formats = "MP3, M4A, M4B, AAC, OGG, FLAC, WAV"
-            )
-            
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            FormatRow(
-                icon = Icons.Default.VideoLibrary,
-                label = "Video",
-                formats = "MP4, MKV, AVI, MOV, WEBM"
-            )
-
-            Spacer(modifier = Modifier.height(12.dp))
-            Text(
-                text = "Unsupported files are skipped during scan. If import fails, verify the file is readable and you granted storage access.",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-}
 
 @Composable
 private fun FormatRow(
@@ -715,151 +553,6 @@ private fun MediaTypeChip(
     )
 }
 
-@Composable
-private fun ScannedFileItem(
-    file: ScannedFile,
-    isSelected: Boolean,
-    onToggleSelect: () -> Unit,
-    onFetchMetadata: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 4.dp)
-            .clickable(onClick = onToggleSelect),
-        colors = CardDefaults.cardColors(
-            containerColor = if (isSelected) 
-                MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
-            else 
-                MaterialTheme.colorScheme.surface
-        )
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Checkbox
-            Checkbox(
-                checked = isSelected,
-                onCheckedChange = { onToggleSelect() }
-            )
-            
-            // Cover/icon
-            Box(
-                modifier = Modifier
-                    .size(56.dp)
-                    .clip(RoundedCornerShape(4.dp))
-                    .background(MaterialTheme.colorScheme.surfaceVariant),
-                contentAlignment = Alignment.Center
-            ) {
-                if (file.metadata?.coverUrl != null) {
-                    AsyncImage(
-                        model = file.metadata.coverUrl,
-                        contentDescription = null,
-                        modifier = Modifier.fillMaxSize(),
-                        contentScale = ContentScale.Crop
-                    )
-                } else {
-                    Icon(
-                        imageVector = file.type.icon,
-                        contentDescription = null,
-                        modifier = Modifier.size(28.dp),
-                        tint = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-            
-            Spacer(modifier = Modifier.width(12.dp))
-            
-            // File info
-            Column(
-                modifier = Modifier.weight(1f)
-            ) {
-                Text(
-                    text = file.metadata?.title ?: file.name,
-                    style = MaterialTheme.typography.bodyLarge,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                
-                // Show authors for books, artist for audio
-                val secondaryText = when (file.type) {
-                    ScannedFileType.MUSIC, ScannedFileType.AUDIOBOOK -> {
-                        buildString {
-                            file.metadata?.authors?.firstOrNull()?.let { append(it) }
-                            file.metadata?.album?.let { 
-                                if (isNotEmpty()) append(" • ")
-                                append(it)
-                            }
-                        }.ifEmpty { null }
-                    }
-                    else -> {
-                        file.metadata?.authors?.takeIf { it.isNotEmpty() }?.joinToString(", ")
-                    }
-                }
-                
-                if (!secondaryText.isNullOrBlank()) {
-                    Text(
-                        text = secondaryText,
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                }
-                
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    // File type badge
-                    Surface(
-                        shape = RoundedCornerShape(4.dp),
-                        color = file.type.color.copy(alpha = 0.2f)
-                    ) {
-                        Text(
-                            text = file.extension.uppercase(),
-                            style = MaterialTheme.typography.labelSmall,
-                            color = file.type.color,
-                            modifier = Modifier.padding(horizontal = 6.dp, vertical = 2.dp)
-                        )
-                    }
-                    
-                    // File size
-                    Text(
-                        text = file.sizeFormatted,
-                        style = MaterialTheme.typography.labelSmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                    
-                    // Metadata status
-                    if (file.metadata != null) {
-                        Icon(
-                            Icons.Default.CheckCircle,
-                            contentDescription = "Metadata loaded",
-                            modifier = Modifier.size(14.dp),
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                }
-            }
-            
-            // Fetch metadata button - for books and audio without metadata
-            if ((file.type == ScannedFileType.BOOK || file.type == ScannedFileType.MUSIC || file.type == ScannedFileType.AUDIOBOOK) && file.metadata == null) {
-                IconButton(onClick = onFetchMetadata) {
-                    Icon(
-                        Icons.Default.Search,
-                        contentDescription = "Fetch metadata",
-                        tint = MaterialTheme.colorScheme.primary
-                    )
-                }
-            }
-        }
-    }
-}
 
 // Data classes
 enum class ScannedFileType(val icon: ImageVector, val color: Color) {
